# agent/pkg/crane/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "crane server API docs. Scope: container management on remote nodes.",
        "title": "crane API Swagger",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/v1",
    "paths": {
        "/containers/{containerPreName}/{containerName}": {
            "delete": {
                "description": "Delete running container by name with status and state",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runtime"
                ],
                "summary": "Delete running container by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "containerName",
                        "name": "containerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "containerPreName",
                        "name": "containerPreName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/containers/{containerPreName}/{containerName}/inspect": {
            "get": {
                "description": "Inspect an existing container by name",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runtime"
                ],
                "summary": "Inspect the container by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "containerName",
                        "name": "containerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "containerPreName",
                        "name": "containerPreName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ContainerJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/containers/{containerPreName}/{containerName}/logs": {
            "get": {
                "description": "Get the logs of an existing container by name",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runtime"
                ],
                "summary": "Get the logs of the container by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "containerName",
                        "name": "containerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "containerPreName",
                        "name": "containerPreName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "paginationSkip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "paginationTake",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/containers/{containerPreName}/{containerName}/status": {
            "get": {
                "description": "Get running container by name with status and state",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runtime"
                ],
                "summary": "Get running containers by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "containerName",
                        "name": "containerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "containerPreName",
                        "name": "containerPreName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ContainerStatusResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "container.Config": {
            "type": "object",
            "properties": {
                "argsEscaped": {
                    "description": "True if command is already escaped (meaning treat as a command line) (Windows specific).",
                    "type": "boolean"
                },
                "attachStderr": {
                    "description": "Attach the standard error",
                    "type": "boolean"
                },
                "attachStdin": {
                    "description": "Attach the standard input, makes possible user interaction",
                    "type": "boolean"
                },
                "attachStdout": {
                    "description": "Attach the standard output",
                    "type": "boolean"
                },
                "cmd": {
                    "description": "Applicable to UNIX platforms",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "domainname": {
                    "description": "Domainname",
                    "type": "string"
                },
                "entrypoint": {
                    "description": "Applicable to UNIX platforms",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "env": {
                    "description": "List of environment variable to set in the container",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "exposedPorts": {
                    "description": "List of exposed ports",
                    "$ref": "#/definitions/nat.PortSet"
                },
                "healthcheck": {
                    "description": "Healthcheck describes how to check the container is healthy",
                    "$ref": "#/definitions/container.HealthConfig"
                },
                "hostname": {
                    "description": "Hostname",
                    "type": "string"
                },
                "image": {
                    "description": "Name of the image as it was passed by the operator (e.g. could be symbolic)",
                    "type": "string"
                },
                "labels": {
                    "description": "List of labels set to this container",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "macAddress": {
                    "description": "Mac Address of the container",
                    "type": "string"
                },
                "networkDisabled": {
                    "description": "Is network disabled",
                    "type": "boolean"
                },
                "onBuild": {
                    "description": "ONBUILD metadata that were defined on the image Dockerfile",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "openStdin": {
                    "description": "Open stdin",
                    "type": "boolean"
                },
                "shell": {
                    "description": "Applicable to UNIX platforms",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stdinOnce": {
                    "description": "If true, close stdin after the 1 attached client disconnects.",
                    "type": "boolean"
                },
                "stopSignal": {
                    "description": "Signal to stop a container",
                    "type": "string"
                },
                "stopTimeout": {
                    "description": "Timeout (in seconds) to stop a container",
                    "type": "integer"
                },
                "tty": {
                    "description": "Attach standard streams to a tty, including stdin if it is not closed.",
                    "type": "boolean"
                },
                "user": {
                    "description": "User that will run the command(s) inside the container, also support user:group",
                    "type": "string"
                },
                "volumes": {
                    "description": "List of volumes (mounts) used for the container",
                    "type": "object",
                    "additionalProperties": {
                        "type": "object"
                    }
                },
                "workingDir": {
                    "description": "Current directory (PWD) in the command will be launched",
                    "type": "string"
                }
            }
        },
...
        "container.HealthConfig": {
            "type": "object",
            "properties": {
                "interval": {
                    "description": "Zero means to inherit. Durations are expressed as integer nanoseconds.",
                    "type": "integer"
                },
                "retries": {
                    "description": "Retries is the number of consecutive failures needed to consider a container as unhealthy.\nZero means inherit.",
                    "type": "integer"
                },
                "startPeriod": {
                    "description": "Zero means to inherit. Durations are expressed as integer nanoseconds.",
                    "type": "integer"
                },
                "test": {
                    "description": "Test is the test to perform to check that the container is healthy.\nAn empty slice means to inherit the default.\nThe options are:\n{} : inherit healthcheck\n{\"NONE\"} : disable healthcheck\n{\"CMD\", args...} : exec arguments directly\n{\"CMD-SHELL\", command} : run command with system's default shell",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "description": "Zero means to inherit. Durations are expressed as integer nanoseconds.",
                    "type": "integer"
                }
            }
        },
...
        "model.Error": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Error"
                    }
                }
            }
        },
...
        "nat.PortMap": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "$ref": "#/definitions/nat.PortBinding"
                }
            }
        },
        "nat.PortSet": {
            "type": "object",
            "additionalProperties": {
                "type": "object"
            }
        },
        "network.Address": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "prefixLen": {
                    "type": "integer"
                }
            }
        },
...
        "network.EndpointSettings": {
            "type": "object",
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "driverOpts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "endpointID": {
                    "type": "string"
                },
                "gateway": {
                    "type": "string"
                },
                "globalIPv6Address": {
                    "type": "string"
                },
                "globalIPv6PrefixLen": {
                    "type": "integer"
                },
                "ipaddress": {
                    "type": "string"
                },
                "ipamconfig": {
                    "description": "Configurations",
                    "$ref": "#/definitions/network.EndpointIPAMConfig"
                },
                "ipprefixLen": {
                    "type": "integer"
                },
                "ipv6Gateway": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "macAddress": {
                    "type": "string"
                },
                "networkID": {
                    "description": "Operational data",
                    "type": "string"
                }
            }
        },
...
        "types.ContainerJSON": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "string"
                },
                "appArmorProfile": {
                    "type": "string"
                },
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "config": {
                    "$ref": "#/definitions/container.Config"
                },
                "created": {
                    "type": "string"
                },
                "driver": {
                    "type": "string"
                },
                "execIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "graphDriver": {
                    "$ref": "#/definitions/types.GraphDriverData"
                },
                "hostConfig": {
                    "$ref": "#/definitions/container.HostConfig"
                },
                "hostnamePath": {
                    "type": "string"
                },
                "hostsPath": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "logPath": {
                    "type": "string"
                },
                "mountLabel": {
                    "type": "string"
                },
                "mounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.MountPoint"
                    }
                },
                "name": {
                    "type": "string"
                },
                "networkSettings": {
                    "$ref": "#/definitions/types.NetworkSettings"
                },
                "node": {
                    "description": "Node is only propagated by Docker Swarm standalone API",
                    "$ref": "#/definitions/types.ContainerNode"
                },
                "path": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "processLabel": {
                    "type": "string"
                },
                "resolvConfPath": {
                    "type": "string"
                },
                "restartCount": {
                    "type": "integer"
                },
                "sizeRootFs": {
                    "type": "integer"
                },
                "sizeRw": {
                    "type": "integer"
                },
                "state": {
                    "$ref": "#/definitions/types.ContainerState"
                }
            }
        },
        "types.ContainerNode": {
            "type": "object",
            "properties": {
                "IP": {
                    "type": "string"
                },
                "addr": {
                    "type": "string"
                },
                "cpus": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "memory": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.ContainerState": {
            "type": "object",
            "properties": {
                "dead": {
                    "type": "boolean"
                },
                "error": {
                    "type": "string"
                },
                "exitCode": {
                    "type": "integer"
                },
                "finishedAt": {
                    "type": "string"
                },
                "health": {
                    "$ref": "#/definitions/types.Health"
                },
                "oomkilled": {
                    "type": "boolean"
                },
                "paused": {
                    "type": "boolean"
                },
                "pid": {
                    "type": "integer"
                },
                "restarting": {
                    "type": "boolean"
                },
                "running": {
                    "type": "boolean"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "description": "String representation of the container state. Can be one of \"created\", \"running\", \"paused\", \"restarting\", \"removing\", \"exited\", or \"dead\"",
                    "type": "string"
                }
            }
        },
        "types.GraphDriverData": {
            "type": "object",
            "properties": {
                "Data": {
                    "description": "data\nRequired: true",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "Name": {
                    "description": "name\nRequired: true",
                    "type": "string"
                }
            }
        },
        "types.Health": {
            "type": "object",
            "properties": {
                "failingStreak": {
                    "description": "FailingStreak is the number of consecutive failures",
                    "type": "integer"
                },
                "log": {
                    "description": "Log contains the last few results (oldest first)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.HealthcheckResult"
                    }
                },
                "status": {
                    "description": "Status is one of Starting, Healthy or Unhealthy",
                    "type": "string"
                }
            }
        },
...
        "types.MountPoint": {
            "type": "object",
            "properties": {
                "destination": {
                    "type": "string"
                },
                "driver": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "propagation": {
                    "type": "string"
                },
                "rw": {
                    "type": "boolean"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.NetworkSettings": {
            "type": "object",
            "properties": {
                "bridge": {
                    "description": "Bridge is the Bridge name the network uses(e.g. `docker0`)",
                    "type": "string"
                },
                "endpointID": {
                    "description": "EndpointID uniquely represents a service endpoint in a Sandbox",
                    "type": "string"
                },
                "gateway": {
                    "description": "Gateway holds the gateway address for the network",
                    "type": "string"
                },
                "globalIPv6Address": {
                    "description": "GlobalIPv6Address holds network's global IPv6 address",
                    "type": "string"
                },
                "globalIPv6PrefixLen": {
                    "description": "GlobalIPv6PrefixLen represents mask length of network's global IPv6 address",
                    "type": "integer"
                },
                "hairpinMode": {
                    "description": "HairpinMode specifies if hairpin NAT should be enabled on the virtual interface",
                    "type": "boolean"
                },
                "ipaddress": {
                    "description": "IPAddress holds the IPv4 address for the network",
                    "type": "string"
                },
                "ipprefixLen": {
                    "description": "IPPrefixLen represents mask length of network's IPv4 address",
                    "type": "integer"
                },
                "ipv6Gateway": {
                    "description": "IPv6Gateway holds gateway address specific for IPv6",
                    "type": "string"
                },
                "linkLocalIPv6Address": {
                    "description": "LinkLocalIPv6Address is an IPv6 unicast address using the link-local prefix",
                    "type": "string"
                },
                "linkLocalIPv6PrefixLen": {
                    "description": "LinkLocalIPv6PrefixLen is the prefix length of an IPv6 unicast address",
                    "type": "integer"
                },
                "macAddress": {
                    "description": "MacAddress holds the MAC address for the network",
                    "type": "string"
                },
                "networks": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/network.EndpointSettings"
                    }
                },
                "ports": {
                    "description": "Ports is a collection of PortBinding indexed by Port",
                    "$ref": "#/definitions/nat.PortMap"
                },
                "sandboxID": {
                    "description": "SandboxID uniquely represents a container's network stack",
                    "type": "string"
                },
                "sandboxKey": {
                    "description": "SandboxKey identifies the sandbox",
                    "type": "string"
                },
                "secondaryIPAddresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.Address"
                    }
                },
                "secondaryIPv6Addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.Address"
                    }
                }
            }
        },
...
        "v1.ContainerStatusResponse": {
            "type": "object",
            "required": [
                "repository",
                "state",
                "status",
                "tag"
            ],
            "properties": {
                "repository": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
...
    }
}